Add support for dynamic fields (columns).
These can be defined at run time, when the column names cannot be known
in advance, for example because they depend on a database of countries or
the administrator's configured language settings.

You'll need to override get_dynamic_fields() in your Resource class to
return a list of Fields; and these will probably need to be instances of
a custom Field class that you define, because the standard classes won't
know how to generate or store the values for fields that are not model
fields, and if they were model fields then you wouldn't need them to be
dynamic.