Remove deprecated providing_args argument to Signal()
This was deprecated in Django 3.1 and will be removed in Django 4.0:

django.utils.deprecation.RemovedInDjango40Warning: The providing_args
argument is deprecated. As it is purely documentational, it has no
replacement. If you rely on this argument as documentation, you can move
the text to a code comment or docstring.

As the text explains it can be removed safely without replacement.