Raise a ValueError for DecimalField parsing
The error-handling code elsewhere (i.e. in resource.py) presupposes that problems
during clean() result in ValueError being raised. For more general exceptions, no
pretty box is shown but instead the exception is shown as a traceback with no means
of understanding what field in particular triggered it. If the DecimalWidget plays
nicely and raises a ValueError for invalid Decimals this problem is fixed