Allow custom change_list_template in admin views using mixins
This allows users to customize the change list templates when using mixins
provided by the library. In particular this enables adding more items to
the object tools, in addition to the ones added by the admin mixins. This
then makes this library compatible with other libraries that extend the
object tools, such as `django-admin-sortable2` or `reversion`.

Fixes #1450