Allow custom change_list_template in admin views using mixins (#1483)
* Allow custom change_list_template in admin views using mixins

This allows users to customize the change list templates when using mixins
provided by the library. In particular this enables adding more items to
the object tools, in addition to the ones added by the admin mixins. This
then makes this library compatible with other libraries that extend the
object tools, such as `django-admin-sortable2` or `reversion`.

Fixes #1450

* Turn custom test change list item into real link

Looks less confusing like this in the demo app.

* Don't make custom change list template test affect the rendering

Should avoid confusion by users using the sample application.

* Update changelog.rst

* Describe potential breaking changes

* Make sure test template customization doesn't render

* corrected typo