Chore: improve logging for debugging purposes (#1991)
* Chore: improve logging for debugging purposes

When tryin to fix tests or see why the import does not work, i usually get a cryptic error message not pointing to the part of the code that failed. It just gives the exception but not where it comes from. By adding exc_info=True, the traceback is visible and its much easier to debug. Also i think warning level should be used instead of debug for these import errors

* Update resources.py